Use the existing karaoke parser

+++ Holo/Media/Providers/ISourceProvider.cs
@@ -174,7 +174,7 @@ public interface ISourceProvider
         AudioVisualizationStyle style,
         long eventStart,
         long eventEnd,
-        double* syllableDurations,
+        long* syllableDurations,
         int syllableDurationsLength,
         int selectedEventIndex
     );
