+++ Mizuki/src/mizuki.zig
@@ -217,7 +217,7 @@ pub export fn GetVisualizationFrameSyllablesView(
     style: c_int,
     event_start: i64,
     event_end: i64,
-    syl_durs: [*]f64,
+    syl_durs: [*]i64,
     syl_durs_len: usize,
 ) ?*frames.Bitmap {
     return buffers.ProcVisualizationFrameSyllablesView(
